% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptorium-package.R
\docType{package}
\name{scriptorium-package}
\alias{scriptorium}
\alias{scriptorium-package}
\title{scriptorium: Easily Install and Load the 'Vellum' Graphics Ecosystem}
\description{
The 'vellum' ecosystem is a set of packages for drawing in R that share a scene model and a common design: 'vellum' (a low-level graphics backend with a Rust engine), 'quill' (a pipe-first grammar of graphics on top of it), and 'gloss' (client-side interactive HTML widgets for the scenes they produce). 'scriptorium' is a meta-package that installs and loads them together, in the spirit of the 'tidyverse'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://schochastics.github.io/scriptorium/}
  \item \url{https://github.com/schochastics/scriptorium}
  \item Report bugs at \url{https://github.com/schochastics/scriptorium/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david.schoch@cynkra.com}

Authors:
\itemize{
  \item David Schoch \email{david.schoch@cynkra.com}
}

}
\keyword{internal}
