% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vellumverse-package.R
\docType{package}
\name{vellumverse-package}
\alias{vellumverse}
\alias{vellumverse-package}
\title{vellumverse: Easily Install and Load the 'Vellum' Graphics Ecosystem}
\description{
The 'vellum' ecosystem is a set of packages for drawing in R that share a scene model and a common design: 'vellum' (a low-level graphics backend with a Rust engine), 'vellumplot' (a pipe-first grammar of graphics on top of it), and 'vellumwidget' (client-side interactive HTML widgets for the scenes they produce). 'vellumverse' is a meta-package that installs and loads them together, in the spirit of the 'tidyverse'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-vellum.github.io/vellumverse/}
  \item \url{https://github.com/r-vellum/vellumverse}
  \item Report bugs at \url{https://github.com/r-vellum/vellumverse/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david.schoch@cynkra.com}

Authors:
\itemize{
  \item David Schoch \email{david.schoch@cynkra.com}
}

}
\keyword{internal}
